+++ packages/vite-plugin-vui/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @fastkit/vite-plugin-vui
 
+## 2.0.42
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fastkit/vui@0.19.42
+
 ## 2.0.41
 
 ### Patch Changes