+++ packages/vui-wysiwyg/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @fastkit/vui-wysiwyg
 
+## 7.0.43
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fastkit/vui@0.19.42
+
 ## 7.0.42
 
 ### Patch Changes